+++ e2e/wallets/intergration-coins/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @coinmasters/integration-coins
 
+## 1.8.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.16
+
 ## 1.8.15
 
 ### Patch Changes